feat: add RowVersion for concurrency control

+++ backend/Versus.API/Models/EloResult.cs
@@ -1,10 +1,10 @@
 namespace Versus.API.Models
 {
     public record EloResult(
-        int WinnerRatingBefore,
-        int WinnerRatingAfter,
-        int WinnerDelta,
-        int LoserRatingBefore,
-        int LoserRatingAfter,
-        int LoserDelta);
+        float WinnerRatingBefore,
+        float WinnerRatingAfter,
+        float WinnerDelta,
+        float LoserRatingBefore,
+        float LoserRatingAfter,
+        float LoserDelta);
 }
